Rolled tests from meterTests.js into meters.js and removed meterTests.js
